+++ sfs-main-extension/src/com/gerantech/towers/sfs/socials/handlers/LobbyModerationHandler.java
@@ -38,7 +38,7 @@ public class LobbyModerationHandler extends BaseClientRequestHandler
         members = lobbyData.getMembers();
         targetMember = null;
         modMember = null;
-        Short action = params.getShort("pr");
+        int action = params.getInt("pr");
         int allSize = members.size();
         for (int i = 0; i < allSize; i++)
         {
@@ -69,14 +69,14 @@ public class LobbyModerationHandler extends BaseClientRequestHandler
 
     private boolean kick(Integer targetId, String targetName)
     {
-        if( modMember.getShort("pr") <= targetMember.getShort("pr") )
+        if( modMember.getInt("pr") <= targetMember.getInt("pr") )
             return false;
 
         User targetUser = lobby.getUserByName(targetId.toString());
         if( targetUser != null )
             getApi().leaveRoom(targetUser, lobby, true, false);
 
-        lobbyRoom.sendComment((short) MessageTypes.M12_COMMENT_KICK, game.player, targetName, (short)-1);// mode = leave
+        lobbyRoom.sendComment( MessageTypes.M12_COMMENT_KICK, game.player, targetName, -1);// mode = leave
         LobbyUtils.getInstance().removeUser(lobbyData, targetId);
         InboxUtils.getInstance().send(MessageTypes.M0_TEXT, "متأسفانه از دهکده اخراج شدی.", BanUtils.SYSTEM_ID, targetId, "");
         OneSignalUtils.getInstance().send(targetName + " متأسفانه " + game.player.nickName + " تو رو از دهکده اخراج کرد.", null, targetId);
@@ -85,11 +85,11 @@ public class LobbyModerationHandler extends BaseClientRequestHandler
 
     private boolean promote(Integer targetId, String targetName)
     {
-        if( modMember.getShort("pr") <= targetMember.getShort("pr") + 1 || targetMember.getShort("pr") >= DefaultPermissionProfile.MODERATOR.getId() )
+        if( modMember.getInt("pr") <= targetMember.getInt("pr") + 1 || targetMember.getInt("pr") >= DefaultPermissionProfile.MODERATOR.getId() )
             return false;
 
-        changePermission((short) (targetMember.getShort("pr") + 1));
-        lobbyRoom.sendComment((short) MessageTypes.M13_COMMENT_PROMOTE, game.player, targetName, targetMember.getShort("pr"));// mode = leave
+        changePermission(targetMember.getInt("pr") + 1);
+        lobbyRoom.sendComment( MessageTypes.M13_COMMENT_PROMOTE, game.player, targetName, targetMember.getInt("pr"));// mode = leave
         InboxUtils.getInstance().send(MessageTypes.M50_URL, "تبریک " + targetName + "، تو توسط " + game.player.nickName + " ریش سپید شدی!", BanUtils.SYSTEM_ID, targetId, "towers://open?controls=tabs&dashTab=3&socialTab=0");
         OneSignalUtils.getInstance().send("تبریک " + targetName + "، تو توسط " + game.player.nickName + " ریش سپید شدی!", null, targetId);
         return true;
@@ -97,22 +97,22 @@ public class LobbyModerationHandler extends BaseClientRequestHandler
 
     private boolean demote(Integer targetId, String targetName)
     {
-        if( modMember.getShort("pr") <= targetMember.getShort("pr") || targetMember.getShort("pr") < DefaultPermissionProfile.STANDARD.getId() )
+        if( modMember.getInt("pr") <= targetMember.getInt("pr") || targetMember.getInt("pr") < DefaultPermissionProfile.STANDARD.getId() )
             return false;
 
-        changePermission((short) (targetMember.getShort("pr") - 1));
-        lobbyRoom.sendComment((short) MessageTypes.M14_COMMENT_DEMOTE, game.player, targetName, targetMember.getShort("pr"));// mode = leave
+        changePermission(targetMember.getInt("pr") - 1);
+        lobbyRoom.sendComment( MessageTypes.M14_COMMENT_DEMOTE, game.player, targetName, targetMember.getInt("pr"));// mode = leave
         //InboxUtils.getInstance().send(MessageTypes.M50_URL, game.player.nickName + " درجه تو رو به سرباز کاهش داد. ", game.player.nickName, game.player.id, targetId, "towers://open?controls=tabs&dashTab=3&socialTab=0");
         //OneSignalUtils.getInstance().send(targetName + "، " + game.player.nickName + " درجه تو رو به سرباز کاهش داد. ", null, targetId);
         return true;
     }
 
-    private void changePermission(short permission)
+    private void changePermission(int permission)
     {
         int index = LobbyUtils.getInstance().getMemberIndex(lobbyData, targetMember.getInt("id"));
         if( index < 0 )
             return;
-        lobbyData.getMembers().getSFSObject(index).putShort("pr", permission);
+        lobbyData.getMembers().getSFSObject(index).putInt("pr", permission);
         LobbyUtils.getInstance().save(lobbyData, null, null,-1,-1,-1, -1, lobbyData.getMembersBytes(), null);
     }
 }
